Rename files and folder structure for QMK pull request
