--- keyboards/Kiwi/readme.md
@@ -1,19 +1,0 @@
-# KiwiKeebsKeebs
-
-![KiwiKeebs](imgur.com image replace me!)
-
-*A short description of the keyboard/project*
-
-* Keyboard Maintainer: [AKiwiKeebs](https://github.com/akiwi92)
-* Hardware Supported: *KiwiKeebs MacroBoard v1.0*
-
-
-Make example for this keyboard (after setting up your build environment):
-
-    make KiwiKeebs:default
-
-Flashing example for this keyboard:
-
-    make KiwiKeebs:default:flash
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
