--- keyboards/kiwikeebs/kb.c
@@ -1,1 +1,0 @@
-#include "kb.h"
