--- keyboards/kiwikeebs/keymaps/default/keymap.c
@@ -1,38 +1,0 @@
-#include "kb.h"
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-	KEYMAP(
-		RESET, KC_2, KC_3,
-		KC_5, KC_6, KC_7, KC_MUTE)
-
-};
-
-const macro_t *action_get_macro(keyrecord_t *record, uint8_t id, uint8_t opt) {
-	keyevent_t event = record->event;
-
-	switch (id) {
-
-	}
-	return MACRO_NONE;
-}
-
-void matrix_init_user(void) {
-}
-
-void matrix_scan_user(void) {
-}
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-	return true;
-}
-
-void encoder_update_user(uint8_t index, bool clockwise) {
-    if (index == 0) { /* First encoder */
-        if (clockwise) {
-            tap_code(KC__VOLUP);
-        } else {
-            tap_code(KC__VOLDOWN);
-        }
-    }
-}
