+++ keyboards/kiwikeebsmacro/config.h
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 AKiwiKeebs92
+Copyright 2020 KiwiKeebs
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -23,8 +23,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define VENDOR_ID    0xFEED
 #define PRODUCT_ID   0x0000
 #define DEVICE_VER   0x0001
-#define MANUFACTURER AKiwiKeebs92
-#define PRODUCT      KiwiKeebsKeebs MacroBoard v.1.0
+#define MANUFACTURER KiwiKeebs
+#define PRODUCT      KiwiKeebs MacroBoard v.1.0
 
 /* key matrix size */
 #define MATRIX_ROWS 2
