+++ keyboards/kiwikeebsmacro/info.json
@@ -1,7 +1,7 @@
 {
-    "keyboard_name": "KiwiKeebs",
+    "keyboard_name": "KiwiKeebs MacroBoard v.10",
     "url": "",
-    "maintainer": "AKiwiKeebs92",
+    "maintainer": "KiwiKeebs",
     "width": 4,
     "height": 2,
     "layouts": {
