+++ keyboards/kiwikeebsmacro/readme.md
@@ -6,10 +6,10 @@ A 6 Key PCB with Rotary Encoder Knob + Button made from solid CNC aluminium.
 
 Keyboard Maintainer: [AKiwi92](https://github.com/akiwi92)
 Hardware Supported: KiwiKeebs MacroBoard V1.0
-Hardware Availability: [1upkeyboards.com](https://www.reddit.com/r/mechmarket/comments/ibijrd/ic_kiwikeebs_6key_macroboard_rotary_encoder_diy/)
+Hardware Availability: [reddit.com](https://www.reddit.com/r/mechmarket/comments/ibijrd/ic_kiwikeebs_6key_macroboard_rotary_encoder_diy/)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make kiwikeebs:default
+    make kiwikeebsmacro:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
